refactor(status): move verifier attribution to read model
