+++ tests/canary/test_maintainability_ratchet.py
@@ -27,7 +27,6 @@ def test_current_repository_debt_is_reviewed_without_line_count_pins() -> None:
     assert report["stale_exception_count"] == 0
     assert set(report["category_counts"]) == {
         "compatibility_facade",
-        "dependency_debt",
         "oversized_decision_function",
     }
     assert report["reviewed_exception_count"] == report["finding_count"]
